Serilog Sentry sink does not load all options from IConfiguration (#290)
The json configuration loading for the Serilog sink was not taking into account all of the options that can be passed to the SentryOptions and SentrySerilogOptions objects.
I have added an additional extension that allows for all the configurable property items that can be loaded from the appsettings.json when NOT using asp.net (such as a windows service).

I added a new method instead of repurposing the existing one to ensure backwards compatibility. I also added an example in the xmldocs so that it is easier to understand how to properly configure the settings for Serilog in json.

This is effectively an extension of the pull request 374 from @rmsy