Unit test for Serilog extensions (#290)
This is likely overkill so please let me know if it should be backed out.
I took what the Serilog group does to convert json strings to IConfiguration objects and piped those through the function in both "zero, one, or many" forms and also with the parameters out of order to ensure it lines up the named properties correctly.