fix: Tweak FlushOnDispose_SendsEnvelope (#4175)
Attempts to fix #2156:
- https://github.com/getsentry/sentry-dotnet/issues/2156

This code makes use of the id returned by `CaptureEvent`. My theory is that might prevent code optimisations resulting in that code running in parallel to the assertion statement.

I also changed the code to call `Dispose` explicitly. I don't think that makes any difference tbh but it does make the method that is being tested more explicit.

#skip-changelog