Experimental SqlLite integration
Relates to #1837:
- https://github.com/getsentry/sentry-dotnet/issues/1837

This is just mucking around for a couple of hours to see how such an integration might be pieced together. If we want to proceed, there is some complexity around the `SqliteDataReader`. This only has an internal constructor, so we'd have to wrap it via a proxy class rather than inheritance....  which would doubtless involve a bunch of boilerplate grunt work.

We'd also have to make decisions about what the hang the spans off. Ideally these DB spans would be children of any active transaction (e.g. the transactions created by the Mvvm integration)... which means storing/retrieving those transaction on the Scope - and as @bruno-garcia recently discovered in the SymbolCollector project, that's fiddly business in GlobalMode (particularly for DB stuff that is likely to be running on a background thread).