fix: Stop warnings from showing in Blazor WASM projects (#4519)
Fix WASM0001 warnings when building Blazor WebAssembly projects that came out from our native bindings.

The fix was to change the way we represent `sentry_value_t` in C# side, instead of using `FieldOffset`, we use a backing field along with getters and setters for the variable as a `double`, or `ulong`.

Fixes #3369